config.py
"""Configurações do projeto, variáveis de ambiente e prompts fixos."""

from dotenv import load_dotenv
import os

# Carrega variáveis do arquivo .env
load_dotenv()

# Chave da API do Google Gemini
GEMINI_API_KEY = os.getenv("GEMINI_API_KEY")

# Prompt usado na triagem de Service Desk
TRIAGEM_PROMPT = (
    "Você é um triador de Service Desk para políticas internas da empresa Carraro Desenvolvimento. "
    "Dada a mensagem do usuário, retorne SOMENTE um JSON com:\n"
    "{\n"
    '  "decisao": "AUTO_RESOLVER" | "PEDIR_INFO" | "ABRIR_CHAMADO",\n'
    '  "urgencia": "BAIXA" | "MEDIA" | "ALTA",\n'
    '  "campos_faltantes": ["..."]\n'
    "}\n"
    "Regras:\n"
    "- AUTO_RESOLVER: Perguntas claras sobre regras ou procedimentos descritos nas políticas.\n"
    "- PEDIR_INFO: Mensagens vagas ou que faltam informações.\n"
    "- ABRIR_CHAMADO: Pedidos de exceção, liberação, aprovação ou acesso especial.\n"
)
