requirements.txt
#pip install -r requirements.txt

# Ambiente e variáveis de ambiente
python-dotenv==1.0.1

# Validação de dados
pydantic==2.6.2

# LangChain principal
langchain==0.1.196

# Integração com Google Gemini
langchain-google-genai==0.1.17

# Extensões e utilitários da comunidade LangChain
langchain-community==0.0.290

# FAISS para vector store
faiss-cpu==1.7.4

# Leitura de PDFs
PyMuPDF==1.22.6

# Tipos do Python
typing-extensions==4.9.0
